Use stage2 boot loader from sdk version 1.0.1
This brings the flash memory layout closer to the 'standard' one used
by cortex-m-rt, with the vector table in front of the text section.

The stage2 boot loader is still first, but that can be considered as a
more or less fixed part of the platform, and not really part of the
firmware.